refactored: the count to size_t and printed unsigned int instead of int to rep the len

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int node_count = 0;
+	size_t node_count = 0;
 
 	while (h != NULL)
 	{
@@ -20,7 +20,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%d] %s", h->len, h->str);
+			printf("[%u] %s", h->len, h->str);
 			printf("\n");
 		}
 		h = h->next;
